Add a default tag
This will add a SHA tag to the docker image by default.

In order to do this we need to introduce a new input `repository` this
input is optional and defaults to
`eu.gcr.io/tradeshift-base/<repo-name>`. The value of the repository
input will be used to make a default tag of the commit SHA.

The new input makes the `registry` input unessesary, as we can easily
determine the registry from the `repository` input. Therefore we also
remove the `registry` input.

As we have some tests in here I wen't ahead and enabled coverage reports
for sonarqube as well.