fix: add support for tags
When we need to build images when a tag is created
with this commit those images will be prefixed with
`tag-<SHA>`. This way we can write ECR lifecycle polices
which will filter images based on tag prefixes in order
to decide how much time or how many images to keep.